Fix Qt6 compiler error

+++ core_lib/src/interface/editor.cpp
@@ -815,7 +815,7 @@ Status Editor::openObject(const QString& strFilePath, const std::function<void(i
     }
     if (!fileInfo.isReadable())
     {
-        dd << QString("Permissions: 0x%1").arg(fileInfo.permissions(), 0, 16);
+        dd << QString("Permissions: 0x%1").arg(static_cast<uint>(fileInfo.permissions()), 0, 16);
         return Status(Status::ERROR_FILE_CANNOT_OPEN,
                       dd,
                       tr("Could not open file"),
